Set a unique clientID for each listener
This allows us to create multiple instances of the listener effectively supporting horizontal scaling. With this comes a
challenge - a message is received in all listener instances. In most use cases, we want a message to be processed only
once. We'll address this by having the listener subscribe to a subject as member of a queue group. NATS will then ensure
that messages are delivered to only one member of the queue group.