Set manualAckMode to true in listener
The default value for manualAckMode is false. With this, a listener automatically acknowledges a message as soon as it's
received. The problem with this is that processing of the message could fail but the NATS streaming server would not
resend the message for a retry. Manually acknowledging the message allows the NATS streaming server to retry processing
of failed messages.