Gracefully shutdown listener
NATS uses heartbeats to determine the health status of its clients. The nats server periodically sends a heartbeat to
each of the clients, gives them time to respond and only marks a client as offline after a certain number of failures to
respond to the heartbeat request. It therefore takes time for a client to be taken offline and messages might in this
time be delivered to an offline client.

Two issues arise from the scenario above:
- Messages delivered to the offline client would only be processed after the nats server realises the client is offline.
  This affects the system's throughput.
- Subsequent messages may be routed to online clients and processed immediately leading to messages being processed out
  of order.

It's ideal if the client closed the connection as soon as its process is closed so that the nats server does not have to
wait for the heartbeat checks to fail.